skye-dialog-manager/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.0.0=pypi_0
aiohttp=3.8.1=pypi_0
aiohttp-cors=0.7.0=pypi_0
aiohttp-jinja2=1.5=pypi_0
aiosignal=1.2.0=pypi_0
altair=4.2.0=pypi_0
antlr4-python3-runtime=4.8=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=pypi_0
asttokens=2.0.5=pyhd8ed1ab_0
async-timeout=4.0.2=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports-zoneinfo=0.2.1=pypi_0
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
base58=2.1.1=pypi_0
beautifulsoup4=4.10.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
blinker=1.4=pypi_0
blis=0.7.8=pypi_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.0.0=pypi_0
catalogue=2.0.8=pypi_0
certifi=2021.10.8=pypi_0
cffi=1.15.0=pypi_0
charset-normalizer=2.0.12=pypi_0
click=8.0.4=py38h06a4308_0
clickclick=20.10.2=pypi_0
colorlog=6.6.0=pypi_0
commonmark=0.9.1=pypi_0
connexion=2.12.0=pypi_0
cudatoolkit=11.1.1=h6406543_10
cymem=2.0.6=pypi_0
data=0.4=pypi_0
debugpy=1.5.1=py38h709712a_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
easydict=1.9=pypi_0
en-core-web-sm=3.4.0=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
et-xmlfile=1.1.0=pypi_0
executing=0.8.3=pyhd8ed1ab_0
faiss-cpu=1.7.2=py3.8_h2a577fa_0_cpu
filelock=3.6.0=pypi_0
flask=2.0.3=pypi_0
frozenlist=1.3.0=pypi_0
fsspec=2022.2.0=pypi_0
funcsigs=1.0.2=pypi_0
future=0.18.2=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
google-auth=2.6.2=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.44.0=pypi_0
huggingface-hub=0.4.0=pypi_0
hydra-colorlog=1.1.0=pypi_0
hydra-core=1.1.1=pypi_0
idna=3.3=pypi_0
ignite=1.1.0=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=5.4.0=pypi_0
inflection=0.5.1=pypi_0
ipykernel=6.9.2=py38h7f3c49e_0
ipython=8.1.1=py38h578d9bd_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
itsdangerous=2.1.1=pypi_0
jedi=0.18.1=py38h578d9bd_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jsonschema=4.4.0=pypi_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.2=py38h578d9bd_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
langcodes=3.3.0=pypi_0
ld_impl_linux-64=2.36.1=hea4e1c9_2
libfaiss=1.7.2=h2bc3f7f_0_cpu
libffi=3.4.2=h7f98852_5
libgcc-ng=11.2.0=h1d223b6_14
libnsl=2.0.0=h7f98852_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=he4da1e4_14
libuv=1.43.0=h7f98852_0
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=13.0.1=he0ac6c6_1
markdown=3.3.6=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h8d4b97c_729
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
multidict=6.0.2=pypi_0
murmurhash=1.0.7=pypi_0
nbclient=0.5.13=pypi_0
nbconvert=6.4.4=pypi_0
nbformat=5.2.0=pypi_0
ncurses=6.3=h9c3ff4c_0
nest-asyncio=1.5.4=pyhd8ed1ab_0
nltk=3.7=pyhd3eb1b0_0
notebook=6.4.10=pypi_0
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
nvidia-ml-py=11.450.51=pypi_0
nvitop=0.5.5=pypi_0
oauthlib=3.2.0=pypi_0
omegaconf=2.1.1=pypi_0
openai=0.22.1=pypi_0
openpyxl=3.0.10=pypi_0
openssl=3.0.0=h7f98852_2
packaging=21.3=pypi_0
pandas=1.4.1=pypi_0
pandas-stubs=1.4.3.220801=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathy=0.6.2=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=9.0.1=pypi_0
pip=22.0.4=pyhd8ed1ab_0
preshed=3.0.6=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.27=pyha770c72_0
protobuf=3.19.4=pypi_0
psutil=5.9.0=py38h497a2fe_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=7.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pydantic=1.9.1=pypi_0
pydeck=0.7.1=pypi_0
pydeprecate=0.3.1=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pympler=1.0.1=pypi_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.1=pypi_0
python=3.8.12=h0744224_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=1.11.0=py3.8_cuda11.1_cudnn8.0.5_0
pytorch-ignite=0.4.8=pypi_0
pytorch-lightning=1.5.10=pypi_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=22.3.0=py38h2035c66_1
readline=8.1=h46c0cb4_0
regex=2022.3.2=pypi_0
requests=2.27.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rich=12.0.0=pypi_0
rsa=4.8=pypi_0
sacremoses=0.0.47=pypi_0
scipy=1.8.0=pypi_0
semver=2.13.0=pypi_0
send2trash=1.8.0=pypi_0
sentencepiece=0.1.96=pypi_0
setproctitle=1.2.2=pypi_0
setuptools=59.5.0=pypi_0
six=1.16.0=pyh6c4a22f_0
smart-open=5.2.1=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.3.1=pypi_0
spacy=3.4.1=pypi_0
spacy-legacy=3.0.9=pypi_0
spacy-loggers=1.0.3=pypi_0
sqlite=3.37.0=h9cd32fc_0
srsly=2.4.4=pypi_0
stack_data=0.2.0=pyhd8ed1ab_0
streamlit=1.7.0=pypi_0
streamlit-chat=0.0.2.1=pypi_0
streamlit-lottie=0.0.3=pypi_0
streamlit-option-menu=0.3.2=pypi_0
swagger-ui-bundle=0.0.9=pypi_0
tbb=2021.5.0=h4bd325d_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.13.3=pypi_0
testpath=0.6.0=pypi_0
thinc=8.1.0=pypi_0
tk=8.6.12=h27826a3_0
tokenizers=0.11.6=pypi_0
toml=0.10.2=pypi_0
toolz=0.11.2=pypi_0
torchmetrics=0.7.2=pypi_0
tornado=6.1=py38h497a2fe_2
tqdm=4.63.0=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
transformers=4.17.0=pypi_0
typer=0.4.2=pypi_0
types-pytz=2022.1.2=pypi_0
typing_extensions=4.1.1=pyha770c72_0
tzdata=2022.1=pypi_0
tzlocal=4.1=pypi_0
urllib3=1.26.8=pypi_0
validators=0.18.2=pypi_0
wasabi=0.10.1=pypi_0
watchdog=2.1.6=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
werkzeug=2.0.3=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
widgetsnbextension=3.6.0=pypi_0
wrapt=1.14.0=pypi_0
xz=5.2.5=h516909a_1
yarl=1.7.2=pypi_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.7.0=pypi_0
zlib=1.2.11=h36c2ea0_1013
